FileExplorer Working with two separate UI elements
Two separate widgets. The TreeView allows you to navigate the filesystem, while the listview below shows the files.

Two separate models are used. One for the treeview, and one for the listview.

A slot on the MainWindow handles the TreeView being clicked. It then uses the QModelIndex of the selected directory to filter the ListView's model (FileModel).

The MainWindow class acts as a type of controller in this situation, connecting the view and model, and handling their events.